backend/requirements.txt
fastapi
uvicorn
httpx
python-multipart
pdfplumber
python-docx
Pillow
tavily-python

